experiments/exp1/baselines/Load_From_S3/requirements.txt
torch
torchvision
numpy
pillow
pandas
boto3